Add barrows minigame (#942)
* Fix double teleport take-off

* Add book read animation

* Add swampy boat and mort myre swamp bridge

* Start adding barrows

* Convert VariableDefinitions.kt to object #839

* Add table column support for vars

* Rename row itemId to rowId

* Add transformed door support

* Add arrow hint packet #543

* Add interface model packet

* Add target check event closes #561

* Fix crypt rat anims

* Door fix

* Add strange old man

* Add barrows door puzzles and more

* Fix spellbook on npc decoding

* Add invalid drain spell message

* Add barrows chest and fix combat defs

* Add rock cavern ranged attack gfx #940

* Fix rock prospecting #940

* Split and fix lrc deposit collapsing #940

* Add barrows integration test

* Formatting

* Stop camera shake after leaving tunnels

* Fix engine unit tests

* Fix drop table max roll override not applying due to table default roll of 1

* Find index immediately when adding npcs to avoid reference issues

* Barrows chest reward fixes

+++ network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/InterfaceOnNpcDecoder.kt
@@ -17,7 +17,7 @@ class InterfaceOnNpcDecoder : Decoder(11) {
         val packed = packet.readInt()
         val npc = packet.readShortLittleEndian().toInt()
         val run = packet.readBooleanAdd()
-        val itemId = packet.readUnsignedShortAdd()
+        val itemId = packet.readUnsignedShortAdd().toShort().toInt()
         return InteractInterfaceNPC(
             npc,
             InterfaceDefinition.id(packed),
